SDL 2 branch: Lots of things.
* Changed SDL.h include from <SDL2/SDL.h> to <SDL.h>. The pkg-config setup
  adds the SDL directory to the include path, not its parent directory.
* Added GTK/SDL2 window embedding.  Note that SDL 2.0.3 (and probably earlier)
  have a bug that makes the whole application freeze when we try to do this,
  so upgrade to the latest version or apply the fix if you're having problems.
* Fixed up input for GTK/SDL2 embedding.
* Restored compatibility with SDL 1.2.

+++ branches/sdl2/src/drivers/sdl/gui.h
@@ -110,7 +110,7 @@ void loadStateFrom();
 void quickLoad();
 void quickSave();
 void changeState(GtkAction *action, GtkRadioAction *current, gpointer data);
-unsigned short GDKToSDLKeyval(int gdk_key);
+int32_t GDKToSDLKeyval(int gdk_key);
 gint convertKeypress(GtkWidget *grab, GdkEventKey *event, gpointer user_data);
 gint handleMouseClick(GtkWidget* widget, GdkEvent *event, gpointer callback_data);
 void handle_resize(GtkWindow* win, GdkEvent* event, gpointer data);
